Making inline code for optics/ directory consistent with others.
The non_inline files should include the header for the file, so that the inline file doesn't need to. This way the inline file has as little in it as possible, meaning inlining into other files comes with less overhead. To avoid no-declaration warnings, the non_inline files now include tmpl_optics.h.