Adding optical power, phase, and transmittance files.
The new rss_ringoccs layout uses structs with far fewer variables. This is to reduce memory consumption, and limit the possibility of memory leaks. This creates a need for functions that can compute common optical variables, like power, phase, and transmittance. These are provided in this commit. The routines are all small enough that they should be inlined. That is, they all make one or two line calls to other functions, like sqrt, polar, atan2, and abs_squared.