Fixing includes for optical transmittance and updating comments.
To reduce the amount of code visible to the compiler with these inline functions, and what is absolutely necessary for the function is included. Functions like sqrt that are the tmpl_math.h are provided in-file as externs. This improves compile time a fair amount. Comments have been added in the top 'doc-string' indicating functions called in the routines.