Clarifying that the portable method does not properly handle signed zero.
The portable version to the absolute value function uses a simple if-then statement. This will not handle negative zero, and |-0| = -0 will occur. Mathematically this is fine, since -0 = 0, but it may cause problems for a user who wishes to differentiate between 0 and -0. Comments have been added specifying this. The non-portable IEEE-754 method handles negative zero correctly.