Removing the 'alt' method of computing the Fresnel kernel. Only the standard method remains.
The 'alt' method existed solely for the sake of being able to run the 'newtonDold' algorithm in rss_ringoccs. It was known from the beginning that newtonDold contained a typo / unjustifiable component, yet miraculously it seemed to work very well. A new algorithm now exists that is far less hand-wavy (contains a solid derivation), and produces better results than newtonDold. Because of this the newtonDold algorithm is being scrapped, and hence these 'alt' functions can be removed too.