Swapping algoritgm to ACM algorithm 116.
The previous algorithm was the GSL algorithm, which used the hypot function to normalize the denominator. The current implementation is modeled on ACM Algorithm 116. It makes the unit tests produce zero error with GCC's _Complex division, my guess is that compiler uses the same algorithm, or something similar. The computational time is similar, compared to the GSL hypot method.