Adding quick divide.
This uses the standard formula for the quotient of two complex numbers. It is susceptible to underflow and overflow since it computes |w|^2 as an intermediate step, but is much faster than the safe method. Most users will not use large numbers where the square will overflow, so offering a quick version is justified.