Blocking the switch-enum warning from clang.
This may be a bug, but it seems impossible to satisfy this warning. If you use a default to catch all remaining values in an enum, you get a warning. If you add all remaining values and omit the default, you get a warning about no default. If you add all types and a default, you get a warning about having a redundant default. I'm not sure if clang intends to fix this, omitting warning for now.